categorize windows.suspicious_threads as malware plugin
